+++ src/render/index.js
@@ -1,38 +1,16 @@
 import { createElement } from '../utils/createElement';
 import DesktopRenderer from '../reconciler/';
-
-let connectToDevTools = () => {};
-
-// from jaredhanson/node-parent-require
-const prequire = function(id) {
-  for (let parent = module.parent; parent; parent = parent.parent) {
-    try {
-      return parent.require(id);
-    } catch (e) {}
-  }
-  throw new Error("Cannot find module '" + id + "' from parent");
-};
-
-try {
-  global.window = global;
-  global.window.WebSocket = prequire('ws');
-  connectToDevTools = prequire('react-devtools-core').connectToDevTools;
-} catch (e) {}
+import connectDevtools from '../devtools';
 
 export let ROOT_NODE = {};
 
 // Renders the input component
 function render(element) {
-  connectToDevTools();
-  DesktopRenderer.injectIntoDevTools({
-    bundleType: 1,
-    version: '0.1.0',
-    rendererPackageName: 'custom-renderer',
-    findHostInstanceByFiber: DesktopRenderer.findHostInstance,
-  });
+  connectDevtools(DesktopRenderer);
 
-  const container = ROOT_NODE;
   ROOT_NODE = createElement('ROOT');
+  const container = ROOT_NODE;
+
   // Returns the current fiber (flushed fiber)
   const node = DesktopRenderer.createContainer(ROOT_NODE);
